move game logic to file in lib
moved game logic from main.rb into gamelogic file in lib folder.
